backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
signalwire
openai
